requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.3.0
click==8.1.3
colorama==0.4.6
coverage==7.2.7
Django==4.2.2
django-cors-headers==4.1.0
django-js-asset==2.1.0
django-mptt==0.14.0
djangorestframework==3.14.0
drf-spectacular==0.26.3
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.11.2
flake8==6.0.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
pluggy==1.2.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyrsistent==0.19.3
pytest==7.4.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
